requirements.txt
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.1
sqlalchemy==2.0.34
psycopg2==2.9.9
plotly
scipy
seaborn
matplotlib
fastparquet
pyarrow
scrapy